feat(core): implement service integration layer phases 1 and 2
Add core types and state management for database connections:
- ConnectionEntry, ConnectionStatus, QueryEvent, SchemaCache structs
- TuskState connect/disconnect/test_connection methods
- QueryService streaming execution with mpsc channels
- Export new types from lib.rs

Update dependencies:
- rusqlite 0.32 -> 0.38 (change Option<u64> to Option<i64> for history)
- dirs 5.0 -> 6.0
- Add futures-util for stream handling

+++ crates/tusk_core/src/models/history.rs
@@ -16,9 +16,9 @@ pub struct QueryHistoryEntry {
     /// The executed SQL
     pub sql: String,
     /// Time to execute (None if not completed)
-    pub execution_time_ms: Option<u64>,
+    pub execution_time_ms: Option<i64>,
     /// Rows returned/affected
-    pub row_count: Option<u64>,
+    pub row_count: Option<i64>,
     /// Error message if query failed
     pub error_message: Option<String>,
     /// Execution timestamp
@@ -32,8 +32,8 @@ impl QueryHistoryEntry {
             id: 0, // Set by database
             connection_id,
             sql: sql.into(),
-            execution_time_ms: Some(result.execution_time_ms),
-            row_count: Some(result.rows.len() as u64),
+            execution_time_ms: Some(result.execution_time_ms as i64),
+            row_count: Some(result.rows.len() as i64),
             error_message: None,
             executed_at: Utc::now(),
         }
